Fix transformResponse
The `transformResponse` fails for invalid JSON. Because it wasn't async, it would always return the promise from the response.json() call. Making it async allowed it to fail the JSON and move on to the next step, but because the body was already read/parsed, the response.text() would always fail. So we have to read the text first, then try and parse it as JSON. If it is JSON, return it. If not, return the text. If reading the text failed, return a response.